+++ ICD.Connect.Partitioning.Commercial/OperationalHoursTelemetryNames.cs
@@ -1,0 +1,15 @@
+﻿namespace ICD.Connect.Partitioning.Commercial
+{
+	public static class OperationalHoursTelemetryNames
+	{
+
+		public const string START_TIME = "StarTime";
+		public const string START_TIME_CHANGED = "StartTimeChanged";
+
+		public const string END_TIME = "EndTime";
+		public const string END_TIME_CHANGED = "EndTimeChanged";
+
+		public const string DAYS = "Days";
+		public const string DAYS_CHANGED = "DaysChanged";
+	}
+}