+++ ICD.Connect.Partitioning.Commercial/Rooms/CommercialRoomConsole.cs
@@ -20,6 +20,7 @@ namespace ICD.Connect.Partitioning.Commercial.Rooms
 			if (instance == null)
 				throw new ArgumentNullException("instance");
 
+			yield return instance.OperationalHours;
 			yield return ConsoleNodeGroup.KeyNodeMap("ConferencePoints", instance.Originators.GetInstancesRecursive<IConferencePoint>(), p => (uint)p.Id);
 			yield return ConsoleNodeGroup.KeyNodeMap("CalendarPoints", instance.Originators.GetInstancesRecursive<ICalendarPoint>(), p => (uint)p.Id);
 			yield return ConsoleNodeGroup.KeyNodeMap("OccupancyPoints", instance.Originators.GetInstancesRecursive<IOccupancyPoint>(), p => (uint)p.Id);